MB-7879: wrap python integers into c_uint64
Using Python values as they are is not safe. We must explicitly
specify uint64_t in order to avoid problems with 32-bit platforms.

All unit tests passed (both 32-bit and 64-bit GNU/Linux).

Reviewed-on: http://review.couchbase.org/26131
Tested-by: Aaron Miller <apage43@ninjawhale.com>

+++ python/couchstore.py
@@ -357,7 +357,7 @@ class CouchStore (object):
         def callback (dbPtr, docInfoPtr, context):
             fn(DocumentInfo._fromStruct(docInfoPtr.contents, self))
             return 0
-        _check(_lib.couchstore_changes_since(self, since, 0, \
+        _check(_lib.couchstore_changes_since(self, c_uint64(since), c_uint64(0), \
                CouchStore.ITERATORFUNC(callback), c_void_p(0)))
 
     def changesSince (self, since):
@@ -379,7 +379,7 @@ class CouchStore (object):
         if endKey:
             ids[1] = SizedBuf(endKey)
             numIDs = 2
-        _check(_lib.couchstore_docinfos_by_id(self, ids, numIDs, 1, \
+        _check(_lib.couchstore_docinfos_by_id(self, ids, numIDs, c_uint64(1), \
                CouchStore.ITERATORFUNC(callback), c_void_p(0)))
 
     @property
